fix name in mpi_test

+++ orangefs-rastropatch/make_bundle.sh
@@ -3,7 +3,12 @@
 make
 make kmod
 
-rm -rf dst
+echo "Clean dst?"
+read yn
+if [ "$yn" = "yes" ]; then
+  rm -rf dst
+fi
+#rm -rf dst
 mkdir dst
 sudo make DESTDIR=`pwd`/dst install
 sudo make DESTDIR=`pwd`/dst kmod_install
